pyproject.toml
[project]
name = "advent-of-code"
version = "0.1.0"
description = "My Advent of Code uhh...code"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
]
